PagesEditor: rename index.jsx into its own PagesEditor component
