+++ app/pages/lab-pages-editor/PagesEditor.jsx
@@ -1,0 +1,40 @@
+// ESLint: don't import global React, and don't use .defaultProps.
+/* eslint-disable no-console */
+/* eslint-disable react/react-in-jsx-scope */
+/* eslint-disable react/require-default-props */
+
+import { StrictMode } from 'react';
+import PropTypes from 'prop-types';
+import DataManager from './DataManager.jsx';
+import Tester from './Tester.jsx';
+
+function PagesEditor(props) {
+  const { params } = props;
+
+  return (
+    <StrictMode>
+      <div>
+        <h5>Pages Editor</h5>
+        <DataManager
+          key={params?.workflowId || '-'} //
+          workflowId={params?.workflowId}
+        >
+          <h6>
+            Workflow
+            {' '}
+            {params?.workflowId}
+          </h6>
+          <Tester />
+        </DataManager>
+      </div>
+    </StrictMode>
+  );
+}
+
+PagesEditor.propTypes = {
+  params: PropTypes.shape({
+    workflowId: PropTypes.string
+  })
+};
+
+export default PagesEditor;
